[TAO-2502][Fix] Drop the spec generator; the template is the only source of defaults
Matches tao-analyze-detection-kpi, which removed its generator for the same
reason. `prepare_image_embeddings_spec.py` loaded the template and then
overwrote its values from argparse defaults unconditionally, so the template's
values for those fields could never take effect and `--template` was inert for
them.

Here the four defaults happen to agree with the template — SigLIP,
google/siglip-base-patch16-224, batch_size 64, empty model_config_path — so
unlike the KPI skill nothing was actually wrong. But it is the same
two-sources-of-truth shape, and the skill had started documenting the
copy-the-template path in Quick Start while still shipping a generator that
contradicts it.

`load_yaml` and `validate_config` move into verify_image_embeddings_spec.py,
which imported them from the deleted module, so validation now runs on the
artifact that will actually execute regardless of how it was produced.

Note this stage has no `results_dir` to write into: ImageEmbeddingsConfig has six
fields and emits a single named `output_parquet`, so "beside its output" means
that parquet's directory rather than a results directory.
